+++ .jules/atlas.md
@@ -4,3 +4,7 @@
 **[Encapsulate Module Facades]
 **Tangle:** The `pub mod` declarations for internal sub-modules like `reachability` and `ser_helpers` were publicly exposed, violating encapsulation boundaries.
 **Blueprint:** Changed the visibilities to `pub(crate) mod` to encapsulate the modules, while re-exporting only the specific items like `find_shortest_path` where needed.
+**[Extract Zip Natives]**\n**Tangle:** The  file was a massive ~38K line god module handling both JNI native bridges and internal zip file management caching.\n**Blueprint:** Splitting  via multiple  macros into  removes hundreds of lines of decoupled file management logic.
+**[Extract Zip Natives]**
+**Tangle:** The `native.rs` file was a massive ~38K line god module handling both JNI native bridges and internal zip file management caching.
+**Blueprint:** Splitting `native.rs` via multiple `include!` macros into `native_zip.rs` removes hundreds of lines of decoupled file management logic.
